Envia holi a 115200 baudios (realmente 117647.1) BRG16 y BRGH añadidos Cambio del segundo argumento del envio de datos seriales Agregada implementacion para enviar un byte y un arreglo de bytes Se cambiaron tabulaciones a espacios en el archivo de constantes main con codigos de prueba del UART

+++ source/SerialDevice.h
@@ -29,7 +29,7 @@ extern "C" {
     // send data
     int
     SerialDevice_SendData(UINT8*    txBuffer,
-                          int       txLength);
+                          UINT8       txLength);
 
     // read data
     int
